Move parametric ops to op_math and make them inherit ControlledOp
