+++ pennylane/ops/qubit/__init__.py
@@ -72,16 +72,11 @@ __ops__ = {
     "RZ",
     "PhaseShift",
     "PCPhase",
-    "ControlledPhaseShift",
     "CPhaseShift00",
     "CPhaseShift01",
     "CPhaseShift10",
     "CPhase",
     "Rot",
-    "CRX",
-    "CRY",
-    "CRZ",
-    "CRot",
     "U1",
     "U2",
     "U3",
